safari-534.53-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.2_White_Space/S7.2_A5_T3-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Parse error
S7.2_A5_T3

PASS Expected parsing failure

TEST COMPLETE
